add -- to su so arguments are correctly parsed
Without this, the --local option intended for gem list is instead
applied to su, which results in a silent failure. Since the su returns
non-zero, Puppet interprets this as needing to run rbenv::gem, which
gets run during every Puppet apply regardless of whether the gem is
already in place or not.